move object generic methods to helpers files
