+++ core/src/utilities/helpers.js
@@ -95,3 +95,16 @@ export const wrapAsPromise = value => {
     resolve(value);
   });
 };
+
+/*
+ * Gets value of the given property on the given object.
+ */
+export const getConfigValueFromObject = (object, property) => {
+  let propIndex = 0;
+  let nextValue = object;
+  const propertyPath = property.split('.');
+  while (nextValue && propIndex < propertyPath.length) {
+    nextValue = nextValue[propertyPath[propIndex++]];
+  }
+  return nextValue;
+};
